refactor: 💡 (BridgeMaker) 변경 위치 model 내로 옮김
